python/s0101/requirements.txt
openai==1.59.3
python-dotenv==1.0.1
beautifulsoup4